W3-02: Convert 22 enum columns to tinyint with PHP enums (#630)
* W3-01: Users table vertical partitioning (phase 1 — dual-write)

Split the 118-column users table into 3 partition tables:
- user_preferences (45 UI/display columns)
- user_activity (13 last_* timestamp columns)
- user_seed_stats (9 seed point/seeding stat columns)

Phase 1 (this PR): dual-write only. All reads still go to users.
The UserPartitionObserver propagates creates/updates/deletes to
the partition tables. Null attributes are skipped so DB defaults
apply for columns not set by the factory/caller.

New files:
- database/migrations/2026_09_07_200000_create_user_partition_tables.php
- database/migrations/2026_09_07_200100_backfill_user_partition_tables.php
- app/Models/UserPreference.php
- app/Models/UserActivity.php
- app/Models/UserSeedStats.php
- app/Observers/UserPartitionObserver.php

Modified:
- app/Models/User.php: added $partitionedColumns map
- app/Models/Traits/HasUserRelationships.php: preference/activity/seedStats HasOne
- app/Providers/AppServiceProvider.php: register observer



* W3-02: Convert 22 enum columns to tinyint with PHP enums

- Add 17 new PHP backed enums (UserStatus, UserPrivacy, UserFontsize,
  UserAcceptPms, UserClickTopic, UserGender, UserTooltip, UserTimeType,
  UserAppendPromotion, AgentAllowMatchType, BitbucketPublic, FaqType,
  LanguageTranslationState, LoginAttemptType, OfferAllowed, OfferVote,
  ReportType, ShoutboxType, SitelogSecurityLevel, TorrentType,
  TorrentCustomFieldType)
- Migration converts all enum columns to TINYINT UNSIGNED using
  add-temp/copy/drop/rename pattern with CASE mapping
- Recreate users(status, added) composite index after column conversion
- Add Eloquent enum casts on all affected models
- Add User::toApiArray() for string-value serialization at API boundary
- Update SecuritySettingsDto to accept both string and int privacy values
- Fix DashboardRepository gender stats to use DB facade (avoids enum cast)
- Fix TorrentUploadRepository to use OfferAllowed enum values
- Fix TorrentUploadController offer query to use enum value
- Fix TorrentSearchRepository showsubcat casts to int
- Fix PageLayoutContext::userFontSize() to convert int to string
- Fix UserDetailRepository::getUser() to use toApiArray()
- Fix UserDetailController status check to use stringValue()
- Update AgentAllowedFamilyTableSeeder to use int matchtype values
- Update tests to use int enum values instead of string literals
- Definition of done: information_schema.columns data_type='enum' = 0

Also reverts W2-04 ViewModel DTO layer per user request:
- Remove app/ViewModels/ directory
- Revert PageServices to return arrays directly
- Revert controllers to call build() without ->toArray()
- Restore RepositorySizeTest baselines



* W3-03: Settings JSON-schema validation

- Add 17 JSON schema files in config/settings-schema/ (one per settings
  prefix: basic, main, smtp, security, authority, tweak, bonus, account,
  torrent, attachment, code, backup, hr, meilisearch, system, permission,
  image_hosting) following JSON Schema draft-07
- Schemas auto-generated from app/Support/Install/settings.default.php
  with permissive types (allowing null and string coercions for DB-stored
  values)
- Add App\Support\SettingsSchemaValidator service that loads and caches
  schemas, validates settings batches against their prefix schema using
  justinrainbow/json-schema
- Add settings:validate artisan command with --prefix filter
- Integrate validation into SettingRepository::saveBatch() — throws
  SettingsValidationException (HTTP 422) on invalid values
- Register SettingsValidationException in Handler for API (Problem Details
  422) and JSON responses
- Add 12 unit tests covering schema loading, validation, command, and
  saveBatch integration
- Fix SettingRepositoryTest to use app() instead of new for DI



* Fix: guard users_status_added_index recreation against duplicate key error

The W3-02 migration unconditionally tried to ADD INDEX users_status_added_index
after converting the status column. On fresh installs (migrate:fresh) the
index already exists from the original create_users_table migration, causing
SQLSTATE 1061 Duplicate key name in CI.

Now checks information_schema.statistics before creating the index.

Fixes all 8 failing CI checks (PHPUnit Unit x4, Code coverage, Octane,
axe-core, k6 performance).



* Fix: relax yes/no enum constraints in settings schemas

CriticalPathTest sets security.iv to empty string '', which failed the
strict enum ["yes","no"] validation. Remove enum constraints from yes/no
fields, keeping them as string|null type only.

Fixes Docker smoke test failure:
SettingsValidationException: Settings validation failed for prefix
'security': security.iv: Does not have a value in the enumeration
["yes","no",null]



* Fix: always drop+recreate users_status_added_index in enum migration

The previous guard only created the index if it didn't exist. On some
MySQL versions, dropping the status column from a composite index leaves
a partial index with only the remaining column(s). This caused
SchemaIndexesTest to fail in Docker smoke test with:

  Index "users_status_added_index" on table "users" has columns [added]
  but expected [status, added]

Now always drops the index (if it exists) and recreates it with the
correct (status, added) column ordering.



* Fix: restore gender='Male' in CriticalPathTest signup

W3-02 changed the test to send gender=0 (integer), but the registration
validation expects string "Male" or "Female". The integer 0 gets cast to
"0", which fromStringSafe maps to UNKNOWN ("N/A"), failing the allowed-
genders check and redirecting back to /signup instead of confirm.php.



---------

+++ app/Repositories/TorrentSearchRepository.php
@@ -60,13 +60,13 @@ class TorrentSearchRepository
         $hasSearchParams = ! empty($searchParams);
         $searchParams['mode'] = $sectiontype;
 
-        $showsubcat = SearchBox::valueWithContext($sectiontype, 'showsubcat'); // whether show subcategory (i.e. sources, codecs) or not
-        $showsource = SearchBox::valueWithContext($sectiontype, 'showsource'); // whether show sources or not
-        $showmedium = SearchBox::valueWithContext($sectiontype, 'showmedium'); // whether show media or not
-        $showcodec = SearchBox::valueWithContext($sectiontype, 'showcodec'); // whether show codecs or not
-        $showstandard = SearchBox::valueWithContext($sectiontype, 'showstandard'); // whether show standards or not
-        $showprocessing = SearchBox::valueWithContext($sectiontype, 'showprocessing'); // whether show processings or not
-        $showaudiocodec = SearchBox::valueWithContext($sectiontype, 'showaudiocodec'); // whether show audio codec or not
+        $showsubcat = (int) SearchBox::valueWithContext($sectiontype, 'showsubcat'); // whether show subcategory (i.e. sources, codecs) or not
+        $showsource = (int) SearchBox::valueWithContext($sectiontype, 'showsource'); // whether show sources or not
+        $showmedium = (int) SearchBox::valueWithContext($sectiontype, 'showmedium'); // whether show media or not
+        $showcodec = (int) SearchBox::valueWithContext($sectiontype, 'showcodec'); // whether show codecs or not
+        $showstandard = (int) SearchBox::valueWithContext($sectiontype, 'showstandard'); // whether show standards or not
+        $showprocessing = (int) SearchBox::valueWithContext($sectiontype, 'showprocessing'); // whether show processings or not
+        $showaudiocodec = (int) SearchBox::valueWithContext($sectiontype, 'showaudiocodec'); // whether show audio codec or not
         $catsperrow = SearchBox::valueWithContext($sectiontype, 'catsperrow'); // show how many cats per line in search box
         $catpadding = SearchBox::valueWithContext($sectiontype, 'catpadding'); // padding space between categories in pixel
